etl-final/connector-service/requirements.txt
Django
djangorestframework
kafka-python
requests
python-dotenv
psycopg2-binary
pymysql
